v2/requirements.txt
pandas
unidecode
postal
dedupe
sentence-transformers
faiss-cpu
transformers
accelerate
